requirements.txt
# Core dependencies
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
SQLAlchemy==2.0.15
psycopg2-binary==2.9.6
celery==5.3.0
redis==4.5.5
diffusers==0.16.1

# AI and ML
torch==2.0.1
transformers==4.29.2
sentence-transformers==2.2.2
accelerate==0.19.0
safetensors==0.3.1
pytorch3d==0.7.4

# 3D processing
numpy==1.24.3
opencv-python==4.7.0.72

# API integrations
requests==2.31.0

# Testing
pytest==7.3.1
pytest-cov==4.0.0
pytest-flask==1.2.0

# Development
black==23.3.0
flake8==6.0.0
mypy==1.3.0
python-dotenv==1.0.0

# Deployment
gunicorn==20.1.0
supervisor==4.2.5

# Additional dependencies
PyJWT==2.7.0
pillow==9.5.0
tqdm==4.65.0
boto3==1.26.137
stripe==5.4.0
